refactor: :recycle: Extracting ChangeAnswers methods into a concern
